refactor(scraper): deepen into ScrapeRun deep module (#1231)
Addresses review on #1231 to fully meet issue #1216 Option 2.

- Rename ScrapeSession -> ScrapeRun (CONTEXT.md reserves Session for
  user-auth); add a ScrapeRun glossary entry distinct from ScrapeReport
  and ScrapeSummary.
- Move orchestration onto the class: prepare, runTheater, runDate,
  loadAvailability, filterDates, finalize, handleFatalError + controlled
  mutators. handleRateLimit/handleDateFailure/summarizeTheater/
  markNotAttempted are now private methods using this.db (no db/session
  clump). index.ts shrinks 862->91 LOC to a thin runScraper entry.
- Remove the five "public for testing only" free exports; tests now drive
  ScrapeRun methods and assert against run.summary.
- Replace native parseInt in createScrapeConfig with a new
  scraper/src/utils/number.ts parseStrictInt; validate SCRAPE_MODE
  against the ScrapeMode union.
- Name the recordError param shape ScrapeErrorEntry (primitive obsession).
- Pass the run itself (a ProgressPublisher) to the strategy instead of
  wrapping emit a second time.
- Fix test DEFAULT_CONFIG to satisfy the full ScrapeConfig type.
- Drop the dead scrapeTheater wrapper; keep loadTheaterMetadata (has a
  dedicated test) as a public export.

tsc --noEmit clean; 254/254 scraper tests pass.
