+++ scraper/src/scraper/scrape-config.ts
@@ -1,4 +1,5 @@
 import type { ScrapeMode } from '../utils/date.js';
+import { parseStrictInt } from '../utils/number.js';
 
 export interface ScrapeConfig {
   movieDelayMs: number;
@@ -7,11 +8,36 @@ export interface ScrapeConfig {
   scrapeDays: number;
 }
 
+const VALID_SCRAPE_MODES: readonly ScrapeMode[] = [
+  'weekly',
+  'from_today',
+  'from_today_limited',
+];
+
+const DEFAULTS = {
+  movieDelayMs: 500,
+  theaterDelayMs: 3000,
+  scrapeMode: 'from_today_limited' as ScrapeMode,
+  scrapeDays: 7,
+};
+
+function envInt(envVar: string, fallback: number): number {
+  const parsed = parseStrictInt(process.env[envVar]);
+  return Number.isFinite(parsed) ? parsed : fallback;
+}
+
+function envScrapeMode(envVar: string, fallback: ScrapeMode): ScrapeMode {
+  const raw = process.env[envVar];
+  return raw && VALID_SCRAPE_MODES.includes(raw as ScrapeMode)
+    ? (raw as ScrapeMode)
+    : fallback;
+}
+
 export function createScrapeConfig(overrides?: Partial<ScrapeConfig>): ScrapeConfig {
   return {
-    movieDelayMs: overrides?.movieDelayMs ?? parseInt(process.env.SCRAPE_MOVIE_DELAY_MS || '500', 10),
-    theaterDelayMs: overrides?.theaterDelayMs ?? parseInt(process.env.SCRAPE_THEATER_DELAY_MS || '3000', 10),
-    scrapeMode: (overrides?.scrapeMode ?? process.env.SCRAPE_MODE as ScrapeMode) ?? 'from_today_limited',
-    scrapeDays: overrides?.scrapeDays || parseInt(process.env.SCRAPE_DAYS || '7', 10),
+    movieDelayMs: overrides?.movieDelayMs ?? envInt('SCRAPE_MOVIE_DELAY_MS', DEFAULTS.movieDelayMs),
+    theaterDelayMs: overrides?.theaterDelayMs ?? envInt('SCRAPE_THEATER_DELAY_MS', DEFAULTS.theaterDelayMs),
+    scrapeMode: overrides?.scrapeMode ?? envScrapeMode('SCRAPE_MODE', DEFAULTS.scrapeMode),
+    scrapeDays: overrides?.scrapeDays ?? envInt('SCRAPE_DAYS', DEFAULTS.scrapeDays),
   };
 }
